M841: drop noop_primitive_operations — redundant .toString() and .toInt()
Eight `StringBuffer.toString()` calls inside string interpolations and
one `.toInt()` on an already-int value:

- formula.dart, asana_csv_importer.dart, enex_page_importer.dart,
  markdown_to_html.dart, roam_page_importer.dart, text_page_importer.dart,
  trello_database_importer.dart, markdown_renderer.dart all dropped
  `${buf.toString()}` → `$buf` (implicit toString in interpolation)
- vault_shell_page.dart's fold dropped the trailing `.toInt()` on a
  sum that's already int; also reformatted the multi-line lambda for
  readability

Info-level lint count 89 → 80.

+++ lib/features/vault/presentation/pages/vault_shell_page.dart
@@ -1346,8 +1346,10 @@ views:
     final pages = await db.select(db.pages).get();
     final databases = await db.select(db.databases).get();
     final relations = await db.select(db.relations).get();
-    final totalBytes =
-        pages.fold<int>(0, (sum, p) => sum + (p.bodyText.length + p.frontmatterJson.length).toInt());
+    final totalBytes = pages.fold<int>(
+      0,
+      (sum, p) => sum + p.bodyText.length + p.frontmatterJson.length,
+    );
     final largest = [...pages]
       ..sort((a, b) => b.bodyText.length.compareTo(a.bodyText.length));
     // Aggregate frontmatter `tags:` lists across pages → frequency map.
